requirements.txt
langchain==0.0.208
deeplake
openai==0.27.8
tiktoken
python-dotenv
pypdf